backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
razorpay==1.4.1
stripe==9.3.0
cloudinary==1.40.0
httpx==0.27.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
pillow==10.3.0
